Facilitate the use of GNU libquadmath if/when needed
Add commented out sections to quickly change to `__float128`.
Template and comment out `miegruneisen.cpp` as it interferes
and is not needed.
Add the ERK(4, 4) file to the project. It might also need
`LIBS += -lquadmath`.
Make `main.cpp` ready to work with 1D equations and
detonations, to be more specific.